requirements.txt
# Healthcare Fraud Detection - Backend Dependencies
# For Vercel Python Serverless Functions

fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.5.3
pydantic==2.11.0
python-dotenv==1.2.1
google-generativeai==0.8.5
slowapi==0.1.9
python-multipart==0.0.21
